requirements.txt
pandas
numpy
spacy
textblob
scikit-learn
SQLAlchemy
feedparser
requests
python-dateutil
APScheduler
python-dotenv
newspaper3k
beautifulsoup4
lxml_html_clean
groq
pydantic
langdetect
rapidfuzz
sentence-transformers
